# Better setup for x86 64 and accelerate f32 distance computation