--- a/wt_sys/Cargo.toml
+++ b/wt_sys/Cargo.toml
@@ -6,5 +6,5 @@ edition = "2021"
 [dependencies]
 
 [build-dependencies]
-bindgen = "0.70.1"
+bindgen = "0.72.1"
 cmake = "0.1.51"