--- a/wt_sys/src/lib.rs
+++ b/wt_sys/src/lib.rs
@@ -1,10 +1,12 @@
 //! Automatically generated bindings for the WiredTiger C library.
 //!
 
+#![allow(clippy::all)]
 #![allow(non_upper_case_globals)]
 #![allow(non_camel_case_types)]
 #![allow(non_snake_case)]
 #![allow(dead_code)]
+#![allow(unnecessary_transmutes)]
 #![allow(rustdoc::invalid_html_tags)]
 #![allow(rustdoc::broken_intra_doc_links)]