Fix some bugs observed in x86 compile locally.

Accelerate f32 distance computation when the input is unaligned bytes using avx512f, otherwise performance is
very poor. Include simsimd hamming distance for scale. Numbers aren't super consistent across runs but this is
still substantially faster than observed on an m4 macbook air, even for hamming distance.

```
f32/coding/l2           time:   [24.146 ns 24.150 ns 24.155 ns]
f32/coding/dot          time:   [18.246 ns 18.247 ns 18.249 ns]
f32/coding/cos          time:   [82.841 ns 83.052 ns 83.427 ns]
f32/l2                  time:   [52.189 ns 52.267 ns 52.413 ns]
f32/dot                 time:   [48.620 ns 48.767 ns 48.998 ns]
f32/cos                 time:   [48.542 ns 48.563 ns 48.601 ns]
i1/hamming              time:   [3.8556 ns 3.8585 ns 3.8620 ns]
```